pyproject.toml
[tool.poetry]
name = "memes_match"
version = "0.1.0"
description = ""
authors = ["Aliaksei Valenta <liverpoolsmile23@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.2"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.0"
PyMySQL = "^1.0.2"
cryptography = "^37.0.3"
Flask-Migrate = "^3.1.0"
connexion = {extras = ["swagger-ui"], version = "^2.14.0"}
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
